back-end/requirements.txt
aniso8601==9.0.1
blinker==1.4
bson==0.5.10
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
configparser==5.0.2
dnspython==1.16.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
Flask-SendGrid==0.7.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pymongo==3.12.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-http-client==3.3.3
pytz==2021.1
requests==2.26.0
sendgrid==6.8.2
six==1.16.0
starkbank-ecdsa==1.1.1
urllib3==1.26.7
Werkzeug==2.0.1
